% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadCopernicusDem.R
\name{downloadCopernicusDem}
\alias{downloadCopernicusDem}
\title{Baixar e processar MDE para uma área de interesse}
\usage{
downloadCopernicusDem(
  roi,
  dir_saida = "../mde",
  n_tentativas = 50,
  n_cores = 18,
  sobrescrever = FALSE,
  recortar_roi = TRUE
)
}
\arguments{
\item{roi}{Região de interesse. Pode ser:
\itemize{
\item Caminho para shapefile (ex: "limites/area.shp")
\item Código de estado brasileiro (ex: "ES", "MG")
\item Objeto sf ou SpatVector
}}

\item{dir_saida}{Diretório onde os dados serão salvos}

\item{n_tentativas}{Número de tentativas de download}

\item{n_cores}{Número de cores para processamento paralelo}

\item{sobrescrever}{Sobrescrever arquivos existentes?}

\item{recortar_roi}{Recortar o MDE final pelos limites exatos da ROI?}
}
\value{
Caminho para o arquivo MDE final
}
\description{
Baixar e processar MDE para uma área de interesse
}
\examples{
# Usando código de estado
processar_mde("ES", "../mde")

# Usando shapefile
processar_mde("limites/bacia.shp", "../mde")

# Usando objeto sf
area_sf <- sf::st_read("minha_area.shp")
processar_mde(area_sf, "../mde")

}
