% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadCopernicusDem.R
\name{downloadCopernicusDemElps}
\alias{downloadCopernicusDemElps}
\title{downloadCopernicusDemElps}
\usage{
downloadCopernicusDemElps(
  aoi,
  outputDir = "./copernicusDem",
  outputFileName = "copernicusDem.tif",
  res = 90,
  type = "DGED",
  outputDirTempFile = "./copernicusDem/tempDirDem",
  keepInvidualTiles = FALSE,
  timeout = 1000,
  ncores = future::availableCores() - 2,
  saveAsInteger = FALSE,
  multiplier = 1,
  showRaster = FALSE,
  retry = 5
)
}
\arguments{
\item{aoi}{sf object. The area of interest for which the DEM is to be downloaded.}

\item{outputDir}{character. The directory where the output files will be saved. Default is './copernicusDem'.}

\item{outputFileName}{character. The name of the output file with a '.tiff' extension. Default is 'copernicusDem.tif'.}

\item{res}{numeric. The desired resolution of the DEM, either 30 or 90 meters. Default is 90.}

\item{type}{character. The type of Copernicus DEM product. Default is 'DGED'.}

\item{outputDirTempFile}{character. The directory for temporary files during the download process. Default is './copernicusDem/tempDirDem'.}

\item{keepInvidualTiles}{logical. Whether to keep the individual downloaded tiles. Default is FALSE.}

\item{timeout}{numeric. The timeout in seconds for each download attempt. Default is 600.}

\item{ncores}{numeric. The number of processor cores to use for parallelizing the download operation. Default is 1 (no parallelization).}

\item{saveAsInteger}{logical. Whether to save the final raster as an integer. Default is FALSE.}

\item{multiplier}{numeric. A multiplier to apply to the data before converting to integer. Default is 1.}

\item{showRaster}{logical. Whether to plot the final raster. Default is FALSE.}

\item{retry}{numeric. The number of retry attempts for failed downloads. Default is 0 (no retries).}
}
\value{
An sf object containing the digital elevation model for the specified area of interest.
}
\description{
Downloads the Copernicus Digital Elevation Model (DEM) for a specified area of interest (AOI).
This function supports parallel processing to expedite the download of individual tiles and includes
error handling to manage any issues that may arise during the download process.
}
\details{
This function downloads the Copernicus Digital Elevation Model (DEM) for a specified area of interest (AOI).
It identifies the necessary tiles based on the AOI, downloads them in parallel (if specified), and merges them into a single DEM file.
The function includes error handling to manage any issues that arise during the download process, ensuring robustness.
}
\examples{
\dontrun{
require(sf)
area_of_interest <- st_read("./area_of_interest.shp")
dem <- downloadCopernicusDemElps(aoi = area_of_interest,
                             outputDir = "./copernicusDem",
                             outputFileName = "copernicusDem.tif",
                             res = 90,
                             type = "DGED",
                             outputDirTempFile = "./tempDirDem",
                             keepInvidualTiles = FALSE,
                             timeout = 600,
                             ncores = 1,
                             saveAsInteger = FALSE,
                             multiplier = 1,
                             showRaster = FALSE,
                             retry = 3)
}

}
\references{
Bielski, C.; López-Vázquez, C.; Grohmann, C.H.; Guth. P.L.; Hawker, L.; Gesch, D.; Trevisani, S.; Herrera-Cruz, V.; Riazanoff, S.; Corseaux, A.; Reuter, H.; Strobl, P., 2024.
Novel approach for ranking DEMs: Copernicus DEM improves one arc second open global topography. IEEE Transactions on Geoscience & Remote Sensing.
https://ieeexplore.ieee.org/document/10440392

European Space Agency, Sinergise (2021). Copernicus Global Digital Elevation Model. Distributed by OpenTopography. https://doi.org/10.5069/G9028PQB. Accessed: 2024-06-27
}
