% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSubBasinsPNRH.R
\name{downloadSubBasinsPNRH}
\alias{downloadSubBasinsPNRH}
\title{Download subbasins of Water Resource National Plan}
\usage{
downloadSubBasinsPNRH(subBasinLevel = 1, outputDir)
}
\arguments{
\item{subBasinLevel}{numeric. subbasin level (1 or 2)}

\item{outputDir}{character. output directory.}
}
\value{
An sf object containing subbasins of Water Resource National Plan
}
\description{
Download subbasins of Water Resource National Plan
}
\details{
File are download to outputdir. Subbasins level 1 and 2 shapefiles are named as "GEOFT_PNRH_SUB1.shp" and "GEOFT_PNRH_SUB2.shp", respectively.
}
\examples{


\dontrun{
require(sf)
area_of_interest <- st_read("./area_of_interest.shp")

pnrh1 <- downloadSubBasinsPNRH(subBasinLevel = 1,
                             output = "./resultados")

}
}
\references{
Nível 1 de Sub-bacias do Plano Nacional de Recursos Hídricos
https://metadados.snirh.gov.br/geonetwork/srv/api/records/f50527b9-24ed-41d5-b063-b5acfb25e10d

Nível 2 de Sub-bacias do Plano Nacional de Recursos Hídricos
https://metadados.snirh.gov.br/geonetwork/srv/api/records/6141f37f-f15d-42e7-8495-ae9ddad0846f
}
