% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_roi.R
\name{load_roi}
\alias{load_roi}
\title{Carregar região de interesse (ROI)}
\usage{
load_roi(roi_input)
}
\arguments{
\item{roi_input}{Pode ser:
\itemize{
\item Caminho para shapefile (character)
\item Objeto sf ou SpatVector
\item Código de estado brasileiro (2 caracteres)
}}
}
\value{
Objeto sf com a geometria da ROI
}
\description{
Carregar região de interesse (ROI)
}
