% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadCopernicusDem.R
\name{downloadCopernicusDem}
\alias{downloadCopernicusDem}
\title{downloadCopernicusDem}
\usage{
downloadCopernicusDem(
  aoi,
  outputDir = "./copernicusDem",
  outputFileName = "copernicusDem.tif",
  res = 90,
  type = "DGED",
  tempDir = "tempDirDem",
  keepInvidualTiles = FALSE
)
}
\arguments{
\item{aoi}{sf object. Area of interest.}

\item{outputDir}{character. Output directory folder path. default = './copernicusDem'.}

\item{outputFileName}{character. Output filename with '.tiff' extension. default = 'copernicusDem.tif'.}

\item{res}{numeric. desired resolution (30 or 90 meters). default = 90.}

\item{type}{character. Copernicus DEM product. Default = 'DGED'.}

\item{tempDir}{character. Temporary directory folder name. Default = 'tempDirDem'.}

\item{keepInvidualTiles}{logical. Keep downloaded tiles?. Default = FALSE.}
}
\value{
Digital elevation model for area of interest
}
\description{
Download Copernicus Digital Elevation Model considering area of interest
}
\examples{

\dontrun{

require(sf)

area_of_interest = st_read("./area_of_interest.shp")


dem = downloadCopernicusDem(aoi = area_of_interest,
                           outputDir = "./copernicusDem",
                           outputFileName = "copernicusDem.tif",
                           res = 90,
                           type = "DGED",
                           tempDir = "tempDirDem",
                           keepInvidualTiles = FALSE)

}


}
