% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadCopernicusDem.R
\name{downloadCopernicusDem}
\alias{downloadCopernicusDem}
\title{downloadCopernicusDem}
\usage{
downloadCopernicusDem(
  aoi,
  outputDir = "./copernicusDem",
  outputFileName = "copernicusDem.tif",
  res = 90,
  type = "DGED",
  outputDirTempFile = "./copernicusDem/tempDirDem",
  keepInvidualTiles = FALSE
)
}
\arguments{
\item{aoi}{sf object. Area of interest.}

\item{outputDir}{character. Output directory folder path. default = './copernicusDem'.}

\item{outputFileName}{character. Output filename with '.tiff' extension. default = 'copernicusDem.tif'.}

\item{res}{numeric. desired resolution (30 or 90 meters). default = 90.}

\item{type}{character. Copernicus DEM product. Default = 'DGED'.}

\item{outputDirTempFile}{character. Temporary directory folder name. Default = 'tempDirDem'.}

\item{keepInvidualTiles}{logical. Keep downloaded tiles?. Default = FALSE.}
}
\value{
Digital elevation model for area of interest
}
\description{
Download Copernicus Digital Elevation Model considering area of interest
}
\details{
Bielski, C.; López-Vázquez, C.; Grohmann, C.H.; Guth. P.L.; Hawker, L.; Gesch, D.; Trevisani, S.; Herrera-Cruz, V.; Riazanoff, S.; Corseaux, A.; Reuter, H.; Strobl, P., 2024.
Novel approach for ranking DEMs: Copernicus DEM improves one arc second open global topography. IEEE Transactions on Geoscience & Remote Sensing.
https://ieeexplore.ieee.org/document/10440392

European Space Agency, Sinergise (2021). Copernicus Global Digital Elevation Model. Distributed by OpenTopography. https://doi.org/10.5069/G9028PQB. Accessed: 2024-06-27
}
\examples{

\dontrun{

require(sf)

area_of_interest = st_read("./area_of_interest.shp")


dem = downloadCopernicusDem(aoi = area_of_interest,
                           outputDir = "./copernicusDem",
                           outputFileName = "copernicusDem.tif",
                           res = 90,
                           type = "DGED",
                           outputDirTempFile = "./tempDirDem",
                           keepInvidualTiles = FALSE)

}


}
